+++ src/partials/misc.js
@@ -138,7 +138,7 @@ module.exports = [
     //     "scope": [
     //         "constant.other.haskell",
     //     ],
-    //     "settings": {"foreground": "#d7ba7d"},
+    //     "settings": {"foreground": "#9cdcfe"},
     // },
     {
         "scope": [
@@ -254,5 +254,28 @@ module.exports = [
     //     ],
     //     "settings": {"foreground": "#9cdcfe"},
     // },
+    // =========================================================================
+    // Smarty
+    // =========================================================================
+    {
+        "scope": [
+            "support.function.built-in.smarty",
+        ],
+        // "settings": { "foreground": "#569cd6" },
+        "settings": { "foreground": "#c586c0" },
+    },
+    {
+        "scope": [
+            "source.smarty punctuation.section.embedded.begin.smarty",
+            "source.smarty punctuation.section.embedded.end.smarty",
+        ],
+        "settings": { "foreground": "#808080" },
+    },
+    {
+        "scope": [
+            "source.smarty",
+        ],
+        "settings": { "foreground": "#4ec9b0" },
+    },
 ];